server/requirements.txt
-i https://pypi.org/simple
alembic==1.12.0
aniso8601==9.0.1
appnope==0.1.3 ; sys_platform == 'darwin'
asttokens==2.4.0
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2 ; python_version >= '3.7'
click==8.1.7
decorator==5.1.1 ; python_version < '3.11' and python_version >= '3.7'
executing==1.2.0
faker==19.6.1
flask==2.3.3
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-migrate==4.0.5
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
gunicorn==21.2.0
honcho==1.1.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipdb==0.13.13
ipython==8.12.2 ; python_version < '3.11' and python_version >= '3.7'
itsdangerous==2.1.2
jedi==0.19.0 ; python_version >= '3.6'
jinja2==3.1.2
mako==1.2.4
markupsafe==2.1.3
matplotlib-inline==0.1.6 ; python_version >= '3.5'
packaging==23.1 ; python_version >= '3.7'
parso==0.8.3 ; python_version >= '3.6'
pexpect==4.8.0 ; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.39 ; python_full_version >= '3.7.0'
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.16.1 ; python_version >= '3.7'
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==68.2.2
six==1.16.0
sqlalchemy==2.0.20
sqlalchemy-serializer==1.4.1
stack-data==0.6.2
tomli==2.0.1 ; python_version < '3.11' and python_version >= '3.7'
traitlets==5.9.0 ; python_version >= '3.7'
typing-extensions==4.7.1 ; python_version >= '3.7'
wcwidth==0.2.6
werkzeug==2.3.7
zipp==3.16.2 ; python_version >= '3.8'
